+++ code/lib/core/i18n/l10n/app_localizations.dart
@@ -1532,6 +1532,18 @@ abstract class AppLocalizations {
   /// In en, this message translates to:
   /// **'Not configured'**
   String get fileOpenBehaviorNotSet;
+
+  /// No description provided for @updateAvailable.
+  ///
+  /// In en, this message translates to:
+  /// **'New version available'**
+  String get updateAvailable;
+
+  /// No description provided for @updateDismiss.
+  ///
+  /// In en, this message translates to:
+  /// **'Dismiss'**
+  String get updateDismiss;
 }
 
 class _AppLocalizationsDelegate