+++ code/lib/core/i18n/l10n/app_localizations_zh.dart
@@ -727,4 +727,10 @@ class AppLocalizationsZh extends AppLocalizations {
 
   @override
   String get fileOpenBehaviorNotSet => '未配置';
+
+  @override
+  String get updateAvailable => '发现新版本';
+
+  @override
+  String get updateDismiss => '忽略';
 }